requirements.txt
numpy
torch
facenet-pytorch
matplotlib
pillow
scikit-learn
pytorch_metric_learning
customtkinter